Split items controller
Split into four controllers

* base (authentication methods) which other api controllers inherit from
* items (show & index items)
* uploads
* downloads
